Enable running on other platforms by refactoring and adding preprocessor directives
